Added Query class
