--- .idea/libraries/Maven__org_slf4j_slf4j_api_1_7_30.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.slf4j:slf4j-api:1.7.30">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/slf4j/slf4j-api/1.7.30/slf4j-api-1.7.30-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
