+++ mockMaxCompute/src/main/java/com/communication/DBMessage.java
