+++ mockMaxCompute/src/main/java/com/communication/DBMessageSource.java
@@ -1,14 +1,10 @@
-package com.communication;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.DataOutputStream;
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.net.Socket;
-
-
-import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.Scanner;
 
 
 public class DBMessageSource {
@@ -18,7 +14,7 @@ public class DBMessageSource {
    
     // give args as ip, port 
    
-	public static void main(String args[]) {
+	public static void main(String[] args) {
 		if(args.length >0 ) {
 			if(args[0]!= null) {
 				messageSourceIp = args[0];
@@ -44,8 +40,8 @@ public class DBMessageSource {
 				System.out.println("Starting thread at message source to send messages to publisher");
 				//while (true) {
 					try {
-        				ObjectMapper objMapper = new ObjectMapper();
-						DataOutputStream dos = new DataOutputStream(publisherSocket.getOutputStream());
+//        				ObjectMapper objMapper = new ObjectMapper();
+//						DataOutputStream dos = new DataOutputStream(publisherSocket.getOutputStream());
 
         				
 //        				DBMessage message = new DBMessage(RequestType.READ, "1", "");
@@ -54,17 +50,41 @@ public class DBMessageSource {
 						//Read 10 times and write/update once
 
 						try{
+							Scanner readQueries;
+							File readQueriesFile;
+							Scanner updateQueries;
+							Scanner insertQueries;
+
 							try{
-								Scanner readQueries = new Scanner(new File("tpch-stream.sql")).useDelimiter(";");
+								readQueriesFile = new File("tpch-stream.sql");
+								readQueries = new Scanner(readQueriesFile).useDelimiter(";");
+							}
+							catch (FileNotFoundException e) {
+								System.out.println("An error occurred.");
+								e.printStackTrace();
+								throw new FileNotFoundException();
+
 							}
 							try{
-								Scanner updateQueries = new Scanner(new File("update.tbl"));
+								updateQueries = new Scanner(new File("update.tbl"));
+							}
+							catch (FileNotFoundException e) {
+								System.out.println("An error occurred.");
+								e.printStackTrace();
+								throw new FileNotFoundException();
 							}
 							try{
-								Scanner insertQueries = new Scanner(new File("insert.tbl"))
+								insertQueries = new Scanner(new File("insert.tbl"));
+							}
+							catch (FileNotFoundException e) {
+								System.out.println("An error occurred.");
+								e.printStackTrace();
+
+								throw new FileNotFoundException();
 							}
 
 							int readOffset = 0;
+
 
 							while(true){
 
@@ -76,7 +96,7 @@ public class DBMessageSource {
 								if (readOffset==10)
 									readOffset=0;
 								if (!readQueries.hasNextLine())
-									readQueries = new File("tpch-stream.sql");
+									readQueries = new Scanner(readQueriesFile).useDelimiter(";");
 								if (updateQueries.hasNextLine()){
 									String updateQuery = updateQueries.nextLine();
 									System.out.println(updateQuery);
@@ -100,7 +120,7 @@ public class DBMessageSource {
 							System.out.println("An error occurred.");
 							e.printStackTrace();
 						}
-						}
+
 //						DBMessage message = new DBMessage(RequestType.INSERT, "1", " \"Roopana \" , 35 ");
 //						dos.writeUTF(objMapper.writeValueAsString(message));
 //						System.out.println("Message sent from source to publisher:"+ objMapper.writeValueAsString(message));
