+++ mockMaxCompute/src/main/java/com/communication/RequestType.java
