+++ mockMaxCompute/src/main/java/com/communication/TableName.java
@@ -1,0 +1,2 @@
+package com.communication;public class TableName {
+}
