+++ mockMaxCompute/src/main/java/com/constants/DatabaseConstants.java
@@ -1,0 +1,20 @@
+/**
+ * 
+ */
+package com.constants;
+
+/**
+ * @author nenem
+ *
+ */
+public class DatabaseConstants {
+
+	public static final String USERNAME = "root";
+	public static final String PASSWORD = "root";
+	public static final String STATE_TABLE = "statetable";
+	public static final String COMMA_DELIMETER = ",";
+	public static final String SPACE_DELIMETER = " ";
+	public static final String NODES_COLUMN_NAME = "nodes";
+	public static final String TABLENAME_COLUMN_NAME = "tablename";
+	
+}
