+++ mockMaxCompute/src/main/java/replication/constants/DatabaseConstants.java
@@ -1,0 +1,13 @@
+package replication.constants;
+
+
+/**
+ * @author nenem
+ *
+ */
+public class DatabaseConstants {
+
+	public static final String USERNAME = "root";
+	public static final String PASSWORD = "root";
+	
+}
